third_party/WebKit/LayoutTests/external/wpt/mediacapture-streams/GUM-impossible-constraint.https-expected.txt

This is a testharness.js-based test.
FAIL Tests that setting an impossible constraint in getUserMedia fails assert_equals: An impossible constraint triggers a OverconstrainedError expected "OverconstrainedError" but got "ConstraintNotSatisfiedError"
Harness: the test ran to completion.
